+++ src/main/resources/application.properties
@@ -2,6 +2,7 @@
 server.port=${PORT:8080}
 server.servlet.context-path=/api
 
+
 # Database Configuration
 #spring.datasource.url=jdbc:postgresql://localhost:5432/metro
 #spring.datasource.username=postgres
@@ -15,74 +16,45 @@ spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 
-# Connection Pool Configuration
-spring.datasource.hikari.connection-timeout=20000
-spring.datasource.hikari.minimum-idle=5
-spring.datasource.hikari.maximum-pool-size=20
-spring.datasource.hikari.idle-timeout=300000
-spring.datasource.hikari.max-lifetime=1200000
 
-# JPA Configuration
+# JPA Configuration - CRITICAL CHANGES
 spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=false
-spring.jpa.properties.hibernate.format_sql=false
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
 
-# Timezone Configuration
 spring.jpa.properties.hibernate.jdbc.time_zone=Asia/Colombo
 spring.jackson.time-zone=Asia/Colombo
 user.timezone=Asia/Colombo
+
 
 # JWT Configuration
 jwt.secret=mySecretKey123456789012345678901234567890123456789012345678901234567890
 jwt.expiration=86400000
 
-# Fix circular dependencies
+# CRITICAL: Fix circular dependencies
 spring.main.allow-circular-references=true
 
 # Actuator Configuration
 management.endpoints.web.exposure.include=health,info
-management.endpoint.health.show-details=when-authorized
 
 # Logging Configuration
 logging.level.com.example.met=INFO
 logging.level.org.springframework.security=WARN
 logging.level.org.hibernate.SQL=WARN
-logging.level.org.springframework.mail=DEBUG
 
 # Application Name Configuration
 spring.application.name=employee-management-system
 
-# Email Configuration - Critical Settings
+# Email Configuration
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
-spring.mail.username=${MAIL_USERNAME:isira.aw@gmail.com}
-spring.mail.password=${MAIL_PASSWORD:cedm mgtv ykaq bgcz}
+spring.mail.username=isira.aw@gmail.com
+spring.mail.password=cedm mgtv ykaq bgcz
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
-spring.mail.properties.mail.smtp.starttls.required=true
-spring.mail.properties.mail.smtp.ssl.trust=smtp.gmail.com
-spring.mail.properties.mail.smtp.connectiontimeout=10000
-spring.mail.properties.mail.smtp.timeout=10000
-spring.mail.properties.mail.smtp.writetimeout=10000
 
-# Debug email issues
-spring.mail.properties.mail.debug=false
-
-# Application Email Configuration
-app.mail.from=${MAIL_FROM:noreply@metropolitan.com}
+# Application Configuration
+app.mail.from=noreply@yourcompany.com
 app.mail.password-reset.subject=Password Reset Request - MET System
-app.mail.password-reset-confirmation.subject=Password Reset Confirmation - MET System
-app.company.name=Metropolitan
-app.company.support-email=${SUPPORT_EMAIL:support@metropolitan.com}
-
-# Password Reset Configuration
 app.password-reset.token-expiration-hours=24
-app.frontend.reset-password-url=${FRONTEND_URL:https://metropolitan-d-production.up.railway.app}/reset-password
-
-# Security Configuration
-spring.security.require-ssl=false
-
-# File Upload Configuration
-spring.servlet.multipart.max-file-size=10MB
-spring.servlet.multipart.max-request-size=10MB
+app.frontend.reset-password-url=https://metropolitan-d-production.up.railway.app/reset-password
